feat: add new page component and update product routes to handle capacity as float

+++ app/api/product/[slug]/route.ts
@@ -75,7 +75,7 @@ export const PUT = async (
         price: parseFloat(price),
         achatPrice: parseFloat(achatPrice),
         stock: parseInt(stock),
-        capacite: parseInt(capacite),
+        capacite: parseFloat(capacite),
         voltage: parseInt(voltage),
         marque,
         designationProduit,
