+++ app/api/product/add/route.ts
@@ -51,7 +51,7 @@ export const POST = async (req: NextRequest) => {
         price: parseFloat(price),
         achatPrice: parseFloat(achatPrice),
         stock: parseInt(stock),
-        capacite: parseInt(capacite),
+        capacite: parseFloat(capacite),
         voltage: parseInt(voltage),
         marque: marque.toLowerCase(),
         designationProduit,
